Creating XTide class. Condensing other classes
