--- src/hwm_savedata.cpp
@@ -1,108 +1,0 @@
-//-------------------------------GPL-------------------------------------//
-//
-// MetOcean Viewer - A simple interface for viewing hydrodynamic model data
-// Copyright (C) 2015  Zach Cobell
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-//
-// The name "MetOcean Viewer" is specific to this project and may not be
-// used for projects "forked" or derived from this work.
-//
-//-----------------------------------------------------------------------//
-
-#include <hwm.h>
-#include <QPrinter>
-
-int hwm::saveHWMMap(QString outputFile, QString filter)
-{
-    if(filter=="PDF (*.pdf)")
-    {
-        QPrinter printer(QPrinter::HighResolution);
-        printer.setPageSize(QPrinter::Letter);
-        printer.setResolution(400);
-        printer.setOrientation(QPrinter::Landscape);
-        printer.setOutputFormat(QPrinter::PdfFormat);
-        printer.setOutputFileName(outputFile);
-
-        QPainter painter(&printer);
-        painter.setRenderHint(QPainter::Antialiasing,true);
-        painter.begin(&printer);
-
-        QPixmap renderedMap = this->map->grab();
-        QPixmap mapScaled = renderedMap.scaledToWidth(printer.width());
-        if(mapScaled.height()>printer.height())
-            mapScaled = renderedMap.scaledToHeight(printer.height());
-        int cw = (printer.width()-mapScaled.width())/2;
-        int ch = (printer.height()-mapScaled.height())/2;
-        painter.drawPixmap(cw,ch,mapScaled.width(),mapScaled.height(),mapScaled);
-
-        painter.end();
-    }
-    else if(filter=="JPG (*.jpg *.jpeg)")
-    {
-        QFile output(outputFile);
-        QSize imageSize(this->map->size().width(),this->map->size().height());
-
-        QImage pixmap(imageSize, QImage::Format_ARGB32);
-        pixmap.fill(Qt::white);
-        QPainter imagePainter(&pixmap);
-        imagePainter.setRenderHints(QPainter::Antialiasing | QPainter::TextAntialiasing | QPainter::SmoothPixmapTransform);
-        this->map->render(&imagePainter,QPoint(0,0));
-
-        output.open(QIODevice::WriteOnly);
-        pixmap.save(&output,"JPG",100);
-    }
-
-    return 0;
-}
-
-
-int hwm::saveRegressionPlot(QString outputFile, QString filter)
-{
-
-    if(filter == "PDF (*.pdf)")
-    {
-        QPrinter printer(QPrinter::HighResolution);
-        printer.setPageSize(QPrinter::Letter);
-        printer.setResolution(400);
-        printer.setOrientation(QPrinter::Landscape);
-        printer.setOutputFormat(QPrinter::PdfFormat);
-        printer.setOutputFileName(outputFile);
-
-        QPainter painter(&printer);
-        painter.setRenderHint(QPainter::Antialiasing,true);
-        painter.begin(&printer);
-
-        this->chartView->render(&painter);
-
-        painter.end();
-    }
-    else if(filter == "JPG (*.jpg *.jpeg)")
-    {
-        QFile output(outputFile);
-        QSize imageSize(this->chartView->size().width(),this->chartView->size().height());
-        QRect chartRect(0,0,this->chartView->size().width(),this->chartView->size().height());
-
-        QImage pixmap(imageSize, QImage::Format_ARGB32);
-        pixmap.fill(Qt::white);
-        QPainter imagePainter(&pixmap);
-        imagePainter.setRenderHints(QPainter::Antialiasing | QPainter::TextAntialiasing | QPainter::SmoothPixmapTransform);
-        this->chartView->render(&imagePainter,chartRect);
-
-        output.open(QIODevice::WriteOnly);
-        pixmap.save(&output,"JPG",100);
-    }
-
-    return 0;
-}
