--- src/noaa_getSet.cpp
@@ -1,62 +1,0 @@
-//-------------------------------GPL-------------------------------------//
-//
-// MetOcean Viewer - A simple interface for viewing hydrodynamic model data
-// Copyright (C) 2015  Zach Cobell
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-//
-// The name "MetOcean Viewer" is specific to this project and may not be
-// used for projects "forked" or derived from this work.
-//
-//-----------------------------------------------------------------------//
-#include <noaa.h>
-
-QString noaa::getNOAAErrorString()
-{
-    return this->NOAAErrorString;
-}
-
-int noaa::getLoadedNOAAStation()
-{
-    return this->NOAAMarkerID;
-}
-
-int noaa::getClickedNOAAStation()
-{
-    QString JunkString;
-    double JunkDouble1,JunkDouble2;
-    return getNOAAStation(JunkString,JunkDouble1,JunkDouble2);
-}
-
-int noaa::setNOAAStation()
-{
-    this->NOAAMarkerID = getNOAAStation(this->CurrentNOAAStationName,
-                                        this->CurrentNOAALon,this->CurrentNOAALat);
-    return 0;
-}
-
-int noaa::getNOAAStation(QString &NOAAStationName, double &longitude, double &latitude)
-{
-    QVariant eval = QVariant();
-    this->map->page()->runJavaScript("returnStationID()",[&eval](const QVariant &v){eval = v;});
-    while(eval.isNull())
-        delayM(5);
-    QStringList evalList = eval.toString().split(";");
-
-    NOAAStationName = evalList.value(1).simplified();
-    latitude = evalList.value(3).toDouble();
-    longitude = evalList.value(2).toDouble();
-
-    return evalList.value(0).toInt();
-}
