--- src/usgs_getDataBounds.cpp
@@ -1,40 +1,0 @@
-//-------------------------------GPL-------------------------------------//
-//
-// MetOcean Viewer - A simple interface for viewing hydrodynamic model data
-// Copyright (C) 2015  Zach Cobell
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-//
-// The name "MetOcean Viewer" is specific to this project and may not be
-// used for projects "forked" or derived from this work.
-//
-//-----------------------------------------------------------------------//
-#include <usgs.h>
-
-int usgs::getDataBounds(double &ymin, double &ymax)
-{
-    int j;
-
-    ymin =  999999999.0;
-    ymax = -999999999.0;
-
-    for(j=0;j<this->USGSPlot.length();j++)
-    {
-        if(this->USGSPlot[j].value<ymin)
-            ymin = this->USGSPlot[j].value;
-        if(this->USGSPlot[j].value>ymax)
-            ymax = this->USGSPlot[j].value;
-    }
-    return 0;
-}
