+++ src/xtide.cpp
@@ -1,0 +1,8 @@
+#include "xtide.h"
+
+XTide::XTide(QWebEngineView *inMap, mov_QChartView *inChart, QDateEdit *inStartDateEdit,
+             QDateEdit *inEndDateEdit, QComboBox *inUnits, QStatusBar *inStatusBar,
+             QObject *parent) : QObject(parent)
+{
+
+}
