+++ ui/MetOceanViewer_main.ui
@@ -83,7 +83,7 @@
            </size>
           </property>
           <property name="currentIndex">
-           <number>2</number>
+           <number>0</number>
           </property>
           <widget class="QWidget" name="tab_noaa">
            <attribute name="title">
